Refactor projeto
