--- APOO-P1/main.java
@@ -1,12 +1,0 @@
-public class main {
-
-    // Variável i definida e inicializada
-    static int i = 10; // exemplo de valor para i
-
-    public static void main(String[] args) {
-
-        // Chama o método inicio() da classe Menu (presumindo que a classe Menu existe)
-        Menu Menu = new Menu();
-        Menu.inicio();
-    }
-}
