+++ src/Main.java
@@ -1,5 +1,10 @@
+// Importa tudo do package Menus
+import Menus.*;
+
 public class Main {
     public static void main(String[] args) {
-        System.out.println("Hello, World!");
+        // Chama o método inicio() da classe Menu (presumindo que a classe Menu existe)
+        Menu Menu = new Menu();
+        Menu.inicio();
     }
 }
